--- src/main/java/auth/AuthConfig.java
@@ -1,48 +1,0 @@
-package auth;
-
-import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.boot.jdbc.DataSourceBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.jdbc.core.JdbcTemplate;
-
-import javax.sql.DataSource;
-
-@SpringBootConfiguration
-public class AuthConfig {
-
-    @Bean
-    public LoginController loginController() {
-        return new LoginController(loginService());
-    }
-
-    @Bean
-    public JwtTokenProvider jwtTokenProvider() {
-        return new JwtTokenProvider();
-    }
-
-    @Bean
-    public LoginService loginService() {
-        return new LoginService(memberDetailsService(), jwtTokenProvider());
-    }
-
-    @Bean
-    public MemberDetailsService memberDetailsService() {
-        return new MemberDetailsService(jdbcTemplate());
-    }
-
-    @Bean
-    public JdbcTemplate jdbcTemplate() {
-        return new JdbcTemplate(dataSource());
-    }
-
-    @Bean
-    public DataSource dataSource() {
-        DataSourceBuilder dataSourceBuilder = DataSourceBuilder.create();
-        dataSourceBuilder.driverClassName("org.h2.Driver");
-        dataSourceBuilder.url("jdbc:h2:mem:test");
-        dataSourceBuilder.username("SA");
-        dataSourceBuilder.password("");
-        System.out.println("asdddddddddd");
-        return dataSourceBuilder.build();
-    }
-}
