+++ src/main/java/nextstep/WebMvcConfiguration.java
@@ -2,17 +2,15 @@ package nextstep;
 
 import auth.*;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 import java.util.List;
 
 @Configuration
-@Import(AuthConfig.class)
 public class WebMvcConfiguration implements WebMvcConfigurer {
-    private LoginService loginService;
-    private JwtTokenProvider jwtTokenProvider;
+    private final LoginService loginService;
+    private final JwtTokenProvider jwtTokenProvider;
 
     public WebMvcConfiguration(LoginService loginService, JwtTokenProvider jwtTokenProvider) {
         this.loginService = loginService;
